+++ examples/cookbook/recipe_thread_safe.cpp
@@ -1,0 +1,28 @@
+//////////////////////////////////////////////////////////////////////////////
+// Copyright (c) 2016-19, Lawrence Livermore National Security, LLC and Umpire
+// project contributors. See the COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (MIT)
+//////////////////////////////////////////////////////////////////////////////
+#include "umpire/strategy/DynamicPool.hpp"
+#include "umpire/strategy/ThreadSafeAllocator.hpp"
+
+#include "umpire/Allocator.hpp"
+#include "umpire/ResourceManager.hpp"
+
+int main(int, char**) {
+  auto& rm = umpire::ResourceManager::getInstance();
+
+  auto pool = rm.makeAllocator<umpire::strategy::DynamicPool>(
+      "pool", rm.getAllocator("HOST"));
+
+  auto thread_safe_pool =
+    rm.makeAllocator<umpire::strategy::ThreadSafeAllocator>
+    ("thread_safe_pool", pool);
+
+  auto allocation = thread_safe_pool.allocate(256);
+  thread_safe_pool.deallocate(allocation);
+
+  return 0;
+}
+